[Patch] Move Read/Write APIs to protobuf.go
